Fixed datatype

+++ sign_detection/GTSDB/ActivationMapBoundingBoxes/use_net.py
@@ -296,7 +296,7 @@ def set_average_value(image, val):
     average = np.average(cv2.mean(image)[:3])
     lut = np.array(range(0, 256)) * (val / average)
     lut = lut.clip(0, 255).astype(np.uint8)
-    res = cv2.LUT(src=image, lut=lut)
+    res = cv2.LUT(src=image.astype(np.uint8), lut=lut)
     return res
